Add new sections for Commands, Overview, Personalization, Settings, and Welcome with corresponding functionality
- Implement CommandsSection to display command groups and their states.
- Create OverviewSection for displaying analytics and server status.
- Add PersonalizationSection for bot customization options including nickname, activity, and avatar.
- Introduce SettingsSection for managing server settings such as auto-reply and language preferences.
- Develop WelcomeSection for configuring welcome messages and channels.
- Update theme.css for improved layout in login mode.
