+++ dashboard/index.html
@@ -16,7 +16,8 @@
     <title>Choco Maid - Discord AI Bot</title>
   </head>
   <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+  <div id="root"></div>
+  <!-- Entry now mounts directly from App.jsx (self-mount pattern) -->
+  <script type="module" src="/src/App.jsx"></script>
   </body>
 </html>
