+++ dashboard/public/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+  <script type="module" src="/src/App.jsx"></script>
   </body>
 </html>
