+++ dashboard/src/components/Header.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+// Placeholder header (currently not used). Can be expanded for future navigation or branding.
+export default function Header(){
+  return null;
+}
