+++ dashboard/src/components/LoginView.jsx
@@ -1,0 +1,54 @@
+import React from 'react';
+
+export default function LoginView({ error, authProcessing, oauthMode, setOauthMode, loginForm, setLoginForm, handleLogin, startDiscordLogin }) {
+  const year = new Date().getFullYear();
+  return (
+    <div className="login-viewport">
+      <div className="login-center fade-in">
+        <div className="login-card card-glass">
+          <div className="login-card-inner">
+            <div className="login-hero mb-3">
+              <div className="logo-orb">
+                <img src="/images.jpg" alt="Choco Maid" className="logo-img" onError={(e)=>{ e.currentTarget.style.display='none'; }} />
+                <span className="orb-text">CM</span>
+              </div>
+              <div>
+                <h1 className="login-title mb-1">Choco Maid</h1>
+                <div className="login-subtitle">Smart assistant & auto‑response manager</div>
+              </div>
+            </div>
+            {error && <div className="alert alert-danger py-2 mb-3">{error}</div>}
+            {authProcessing ? <div className="auth-processing vstack gap-3 text-center py-4">
+              <div className="spinner-border text-light mx-auto" style={{width:'2.5rem', height:'2.5rem'}} role="status"><span className="visually-hidden">Loading...</span></div>
+              <div className="small text-muted">Completing Discord sign‑in…</div>
+            </div> : oauthMode ? <div className="vstack gap-3">
+                <p className="text-muted small m-0">Authenticate with your Discord account to access your servers and manage bot settings.</p>
+                <button onClick={startDiscordLogin} className="btn btn-discord-cta">
+                  <span className="ico me-1"><i className="fa-brands fa-discord" /></span>
+                  <span>Login with Discord</span>
+                </button>
+                {!(typeof import.meta!=='undefined' && import.meta.env && import.meta.env.VITE_DISABLE_LEGACY_LOGIN) && <button type="button" className="btn btn-link p-0 small" onClick={()=>setOauthMode(false)}>Use legacy admin login</button>}
+              </div> : <form onSubmit={handleLogin} className="vstack gap-3">
+                <div>
+                  <label className="form-label small mb-1">Username</label>
+                  <input className="form-control" placeholder="admin" value={loginForm.username} onChange={e=>setLoginForm({...loginForm, username:e.target.value})} />
+                </div>
+                <div>
+                  <label className="form-label small mb-1">Password</label>
+                  <input className="form-control" type="password" placeholder="••••••" value={loginForm.password} onChange={e=>setLoginForm({...loginForm, password:e.target.value})} />
+                </div>
+                <div className="d-flex justify-content-between align-items-center mt-1">
+                  <button type="submit" className="btn btn-brand flex-grow-1">
+                    <i className="fa-solid fa-right-to-bracket me-2" />
+                    Login
+                  </button>
+                  <button type="button" className="btn btn-link p-0 small ms-3" onClick={()=>setOauthMode(true)}>Discord login</button>
+                </div>
+              </form>}
+            <div className="login-footer small text-muted mt-4">© {year} Choco Maid • Not affiliated with Discord</div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
